+++ scripts/utils/browser_tools.js
@@ -1,0 +1,15 @@
+module.exports = {
+  isMobile: function(){
+    var IS_MOBILE = (
+      navigator.userAgent.match(/Android/i)
+        || navigator.userAgent.match(/webOS/i)
+        || navigator.userAgent.match(/iPhone/i)
+        || navigator.userAgent.match(/iPad/i)
+        || navigator.userAgent.match(/iPod/i)
+        || navigator.userAgent.match(/BlackBerry/i)
+        || navigator.userAgent.match(/Windows Phone/i)
+    );
+
+    return IS_MOBILE;
+  }
+};
